--- eureka/src/main/java/com/ideasconnections/microservicios/api/eureka/EurekaApplication.java
@@ -1,15 +1,0 @@
-package com.ideasconnections.microservicios.api.eureka;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@EnableEurekaServer
-@SpringBootApplication
-public class EurekaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EurekaApplication.class, args);
-	}
-
-}
