--- eureka/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-#nombre del servidor
-spring.application.name=microservixio-eureka-server
-#puerto
-server.port=8761
-
-#clientes
-eureka.client.register-with-eureka=false
-eureka.client.fetch-registry=false
