--- eureka/src/test/java/com/ideasconnections/microservicios/api/eureka/EurekaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ideasconnections.microservicios.api.eureka;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EurekaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
