+++ microservicios-eureka/src/main/java/com/ideasconnections/microservicios/app/eureka/MicroserviciosEurekaApplication.java
@@ -1,0 +1,15 @@
+package com.ideasconnections.microservicios.app.eureka;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+
+@EnableEurekaServer
+@SpringBootApplication
+public class MicroserviciosEurekaApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MicroserviciosEurekaApplication.class, args);
+	}
+
+}
