+++ microservicios-eureka/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+#nombre del servidor
+spring.application.name = microservicio-eureka-server
+#puerto
+server.port = 8761
+
+#clientes
+eureka.client.register-with-eureka = false
+eureka.client.fetch-registry = false
