+++ microservicios-eureka/src/test/java/com/ideasconnections/microservicios/app/eureka/MicroserviciosEurekaApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ideasconnections.microservicios.app.eureka;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MicroserviciosEurekaApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
