+++ microservicios-usuarios/src/main/java/com/ideasconnections/microservicios/app/usuarios/MicroserviciosUsuariosApplication.java
@@ -1,0 +1,13 @@
+package com.ideasconnections.microservicios.app.usuarios;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MicroserviciosUsuariosApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MicroserviciosUsuariosApplication.class, args);
+	}
+
+}
