+++ microservicios-usuarios/src/main/resources/application.properties
@@ -1,0 +1,15 @@
+#identificar client
+spring.application.name = usuarios
+server.port = ${PORT:0}
+eureka.instance.instance-id = ${spring.application.name}:${ramdom.value}
+
+#ruta comunicacion
+eureka.client.service-url.defaulZone = http://localhost:8761/eureka
+
+spring.datasource.url = jdbc:mysql://localhost:3306/microservicios_examenes
+spring.datasource.username = root
+spring.datasource.password =
+#spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
+spring.jpa.generate-ddl=true
+logging.level.org.hibernate.SQL = debug
