+++ microservicios-usuarios/src/test/java/com/ideasconnections/microservicios/app/usuarios/MicroserviciosUsuariosApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ideasconnections.microservicios.app.usuarios;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MicroserviciosUsuariosApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
