--- usuarios/src/main/java/com/ideasconnections/microservicios/app/usuarios/UsuariosApplication.java
@@ -1,13 +1,0 @@
-package com.ideasconnections.microservicios.app.usuarios;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UsuariosApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UsuariosApplication.class, args);
-	}
-
-}
