--- usuarios/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-
-spring.datasource.url= jbdc:mysql://localhost:3306/db_microservicios_examenes
-spring.datasource.username=root
-spring.datasource.password=
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-spring.jpa.database-platform=org.hibernate.dialect.MySQL5.0.128Dialect
-spring.jpa.generate-ddl=true
-logging.level.org.hibernate.SQL= debug
