--- usuarios/src/test/java/com/ideasconnections/microservicios/app/usuarios/UsuariosApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ideasconnections.microservicios.app.usuarios;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UsuariosApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
